requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.9
python-decouple==3.1
static3==0.7.0
django-ckeditor==5.3.0
django-js-asset==1.1.0
Pillow==5.4.1
djangorestframework==3.6.0
djangorestframework-jwt==1.11.0
Markdown==3.0.1
PyJWT==1.7.1
pytz==2018.9
